Create importer base classes, fix create_importer
The base classes for bids importing for modalities, sessions and
subjects have been created and the base methods have been written.
Specific importers now have to be designed for importing from pia to the
desired format (ucsfbids).

create_importer seemed to be using the `session` parameter regardless of
whether it was creating a session, modality or subject importer. This
was fixed.